Refactored re-usable test utilities into test-utils.c.
* Fixed issues URL in CONTRIBUTING.
* Fixed copyright year in LICENSE.
