+++ src/Makefile.in
@@ -31,7 +31,12 @@ sha2.o: sha2.h memzero.h
 hmac.o: hmac.h sha2.h memzero.h options.h
 pbkdf2.o: pbkdf2.h sha2.h hmac.h memzero.h sha2.h
 
-test: $(libname) test.c
+TEST_OBJS = test.o test-utils.o
+
+test: $(libname) $(TEST_OBJS)
+
+test.o: test-utils.h
+test-utils.o: test-utils.h
 
 .PHONY: check
 check: all
@@ -66,7 +71,7 @@ uninstall:
 clean:
 	rm -rf $(libname)
 	rm -f $(OBJS)
-	rm -f test
+	rm -f test $(TEST_OBJS)
 	rm -rf *.dSYM
 
 .PHONY: distclean
